Format

+++ tests/test_rb_min.py
@@ -11,17 +11,14 @@ def rb_minimised_is_reasonable(molecule):
     coords = opt_rb_coords(
         py_coords=molecule.coordinates,
         py_bonded_matrix=molecule.bond_matrix,
-        py_r0_matrix=np.asarray(molecule.graph.eqm_bond_distance_matrix,
-                                dtype='f8'),
-        py_k_matrix=1 * np.ones((n_atoms, n_atoms),
-                                dtype='f8'),
-        py_c_matrix=0.01 * np.ones((n_atoms, n_atoms),
-                                   dtype='f8'),
-        py_exponent=8
+        py_r0_matrix=np.asarray(molecule.graph.eqm_bond_distance_matrix, dtype="f8"),
+        py_k_matrix=1 * np.ones((n_atoms, n_atoms), dtype="f8"),
+        py_c_matrix=0.01 * np.ones((n_atoms, n_atoms), dtype="f8"),
+        py_exponent=8,
     )
 
     molecule.coordinates = coords
-    molecule.print_xyz_file(filename='tmp.xyz')
+    molecule.print_xyz_file(filename="tmp.xyz")
 
     assert are_coords_reasonable(coords)
 
@@ -29,7 +26,7 @@ def rb_minimised_is_reasonable(molecule):
 def test_h2():
     """Simple 1D optimisation of H2, slightly displaced from its minimum"""
 
-    h2 = Molecule(atoms=[Atom('H'), Atom('H', x=1.1)])
+    h2 = Molecule(atoms=[Atom("H"), Atom("H", x=1.1)])
 
     # Add a bond between the two H atoms
     h2.graph.add_edge(0, 1)
@@ -40,17 +37,18 @@ def test_h2():
 
 def test_alkanes():
 
-    rb_minimised_is_reasonable(molecule=Molecule(smiles='C'))
-    rb_minimised_is_reasonable(molecule=Molecule(smiles='CCCC'))
+    rb_minimised_is_reasonable(molecule=Molecule(smiles="C"))
+    rb_minimised_is_reasonable(molecule=Molecule(smiles="CCCC"))
 
 
 def _test_tmp():
 
-    butane = Molecule(smiles='CCCCCCCC')
+    butane = Molecule(smiles="CCCCCCCC")
 
     from autode.conformers.conf_gen import get_simanl_conformer
+
     tmp = get_simanl_conformer(butane, save_xyz=False)
-    tmp.print_xyz_file(filename='tmp.xyz')
+    tmp.print_xyz_file(filename="tmp.xyz")
 
     # butane.populate_conformers(n_confs=5)
     # for i, conf in enumerate(butane.conformers):
